+++ src/index.tsx
@@ -3,7 +3,6 @@ import ReactDOM from 'react-dom/client';
 import './index.css';
 import App from './App';
 import reportWebVitals from './reportWebVitals';
-import Main from './Main';
 import { BrowserRouter as Router } from 'react-router-dom';
 
 const root = ReactDOM.createRoot(
@@ -14,7 +13,6 @@ root.render(
   <React.StrictMode>
     <Router>
       <App />
-      <Main />
     </Router>
   </React.StrictMode>
 );
